redid using existing model
