+++ jpo-ode-core/src/main/java/us/dot/its/jpo/ode/model/OdeAsn1Metadata.java
@@ -3,21 +3,11 @@ package us.dot.its.jpo.ode.model;
 import java.util.ArrayList;
 import java.util.List;
 
-public class OdeAsn1Metadata extends OdeLogMetadata {
+public class OdeAsn1Metadata extends OdeLogMetadataReceived {
 
    private static final long serialVersionUID = -8601265839394150140L;
 
    private List<Asn1Encoding> encodings = new ArrayList<Asn1Encoding>();
-
-   private ReceivedMessageDetails receivedMessageDetails;
-   
-   public ReceivedMessageDetails getReceivedMessageDetails() {
-      return receivedMessageDetails;
-   }
-
-   public void setReceivedMessageDetails(ReceivedMessageDetails receivedMessageDetails) {
-      this.receivedMessageDetails = receivedMessageDetails;
-   }
 
    public OdeAsn1Metadata() {
       super();
